requirements.txt
data==0.4
Pillow==8.3.2
